Changed sum type from int to double in vector distance calculation

+++ src/StatisticsHelper.java
@@ -1,6 +1,6 @@
 public class StatisticsHelper {
-    public static double calculateDistance(double x[], double y[]){
-        int sum = 0;
+    public static double calculateDistance(double x[], double y[]) {
+        double sum = 0;
         for (int i = 0; i < x.length; i++) {
             sum += Math.pow(x[i] - y[i], 2);
         }
